+++ src/pages/_app.tsx
@@ -1,23 +1,34 @@
 import { FC } from "react";
 import type { AppProps } from "next/app";
+import { ApolloProvider } from "@apollo/client";
 
 import { AuthProvider } from "../auth/useAuth";
 import "../styles/constants.css";
 import "../styles/reset.css";
 import "../styles/globals.css";
+import useApolloClient from "../browser-lib/graphql/useApolloClient";
 import useTheme from "../hooks/useTheme";
 import { SearchProvider } from "../context/SearchContext";
+import ErrorBoundary from "../components/ErrorBoundary";
+import { BookmarksProvider } from "../hooks/useBookmarks";
 
 const MyApp: FC<AppProps> = ({ Component, pageProps }) => {
   useTheme();
+  const apolloClient = useApolloClient();
 
   return (
     <>
-      <AuthProvider>
-        <SearchProvider>
-          <Component {...pageProps} />
-        </SearchProvider>
-      </AuthProvider>
+      <ErrorBoundary>
+        <AuthProvider>
+          <ApolloProvider client={apolloClient}>
+            <BookmarksProvider>
+              <SearchProvider>
+                <Component {...pageProps} />
+              </SearchProvider>
+            </BookmarksProvider>
+          </ApolloProvider>
+        </AuthProvider>
+      </ErrorBoundary>
     </>
   );
 };